+++ tests/test_describe.py
@@ -2,7 +2,11 @@
 # Copyrights licensed under the BSD License. See the accompanying LICENSE
 # file for terms.
 
+import six
 from figgis import Config, Field, ListField
+
+
+STR = six.text_type.__name__
 
 
 def test_flat():
@@ -18,10 +22,10 @@ def test_flat():
     assert desc == """\
 choices (type=int, choices=[1, 2, 3]) - choices
 default (type=int, default=0) - default
-list_ (type=list(str)) - list_
-nodesc (type=str)
-not_nullable (type=str, non-nullable)
-required (type=int, required) - required"""
+list_ (type=list({str})) - list_
+nodesc (type={str})
+not_nullable (type={str}, non-nullable)
+required (type=int, required) - required""".format(str=STR)
 
 
 def test_nested():
@@ -41,10 +45,10 @@ def test_nested():
     assert desc == """\
 config (type=SubConfig)
   config (type=SubSubConfig)
-    value1 (type=str)
-    value2 (type=str)
-  value (type=str)
-value (type=str)"""
+    value1 (type={str})
+    value2 (type={str})
+  value (type={str})
+value (type={str})""".format(str=STR)
 
 
 def test_nested_magic_help():
@@ -67,10 +71,10 @@ def test_nested_magic_help():
     assert desc == """\
 config (type=SubConfig) - subconfig
   config1 (type=SubSubConfig) - override
-    value1 (type=str)
-    value2 (type=str)
+    value1 (type={str})
+    value2 (type={str})
   config2 (type=SubSubConfig) - subsubconfig
-    value1 (type=str)
-    value2 (type=str)
-  value (type=str)
-value (type=str)"""
+    value1 (type={str})
+    value2 (type={str})
+  value (type={str})
+value (type={str})""".format(str=STR)